add skel for avalanche contract deploy erc20
